Split out serializers and models into separate classes
