+++ pyproject.toml
@@ -79,6 +79,8 @@ source = ['requests_cache']
 omit = [
     'requests_cache/__init__.py',
     'requests_cache/backends/__init__.py',
+    'requests_cache/models/__init__.py',
+    'requests_cache/serializers/__init__.py',
 ]
 
 [tool.isort]
