+++ requests_cache/serializers/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa: F401
+from .base import BaseSerializer
+from .pickle import PickleSerializer, SafePickleSerializer
